refactor: extract NFT transactions tab to a separate file
